requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
plotly
shap
matplotlib
openai
python-dotenv
imbalanced-learn